Set up GHDL so we don't always need Vivado access.
Note that the MMU can't be used because Xilinx only provides Verilog
models of their components. Also, some filenames will need to be changed
for the Makefile to work properly as it requires hdl/<component>.vhd and
sim/<component>.vhd for everything. We'll do this once Hugo merges his
changes.